refactor: move BioCardListItem level up
