+++ src/components/GenericPagesComponents/BioCardListItem/index.ts
